general fix of non int slice

+++ src/common/modeling.py
@@ -295,7 +295,7 @@ class LLM(LanguageModel):
                 prompt_length = len(self.model.tokenizer(prompt))
                 self.logger.log(
                     orange(f"INFO: Prompt is too long. Length: {prompt_length}. Truncating it hard with factor 0.9."))
-                prompt = prompt[:0.9 * len(prompt)]
+                prompt = prompt[:int(0.9 * len(prompt))]
             output = self.model(prompt,
                                 eos_token_id=self.model.tokenizer.eos_token_id,
                                 pad_token_id=self.model.tokenizer.pad_token_id,
